pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "links"
version = "0.13.0"
description = "Links: verifiable claim exchange with group policy controls"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "typer>=0.9",
  "pydantic>=2.0",
  "networkx>=3.0",
  "pynacl>=1.5",
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "requests>=2.31",
  "cryptography>=42.0",
  "pywikibot",
]

[project.scripts]
links = "links.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["links"]

